wip create link component

+++ database/migrations/2019_07_01_193235_create_links_table.php
@@ -15,7 +15,7 @@ class CreateLinksTable extends Migration
     {
         Schema::create('links', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('profile_id');
+            $table->unsignedBigInteger('profile_id');
             $table->string('name');
             $table->string('url');
             $table->integer('type');
