+++ database/migrations/2019_07_01_194022_create_projects_table.php
@@ -15,7 +15,7 @@ class CreateProjectsTable extends Migration
     {
         Schema::create('projects', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('profile_id');
+            $table->unsignedBigInteger('profile_id');
 
             $table->foreign('profile_id')->references('id')->on('profiles')->onDelete('cascade');
             $table->timestamps();
